Fixing open/close sidebar bug on mobile and page reload bug when clicking on search button
